# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The backend REST API for Tunes",
        "title": "Tunes backend API",
        "contact": {
            "name": "Jack Gitter",
            "email": "jack.a.gitter@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/posts": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Creates a post for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Creates a post for the current user",
                "parameters": [
                    {
                        "description": "Information required to create a post",
                        "name": "CreatePostDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreatePostDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PostPreview"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/posts/current/{songID}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a specific post for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Get a specific post for the current user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The songID of the posted song",
                        "name": "songID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PostPreview"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Gets the comments of a post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Gets the comments of a post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The songID of the posted song",
                        "name": "songID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The user who posted the post spotify ID",
                        "name": "spotifyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pagination Key. In the form of UTC timestamp",
                        "name": "createdAt",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PaginationResponse-array_responses_Comment-time_Time"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Updates a post made by the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Updates a post made by the current user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The songID of the posted song",
                        "name": "songID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The fields to update",
                        "name": "UpdatePostDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdatePostRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PostPreview"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/posts/dislikes/{spotifyID}/{songID}": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Dislikes a post for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Dislikes a post for the current user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Song ID of the post to dislike",
                        "name": "spotifyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Spotify ID of the user who posted the song",
                        "name": "songID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/posts/likes/{spotifyID}/{songID}": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Likes a post for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Likes a post for the current user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Song ID of the post to like",
                        "name": "spotifyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Spotify ID of the user who posted the song",
                        "name": "songID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/posts/previews/users/current": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all of a users post previews",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Get all of a users post previews",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The user whos posts are recieved. Value is a spotify ID",
                        "name": "spotifyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pagination Key. Format is UTC timestamp",
                        "name": "createdAt",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PostPreview"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/posts/previews/users/{spotifyID}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all of a users post previews",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Get all of a users post previews",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The user whos posts are recieved. Value is a spotify ID",
                        "name": "spotifyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pagination Key. Format is UTC timestamp",
                        "name": "createdAt",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PostPreview"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/posts/votes/current/{posterSpotifyID}/{songID}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Removes a vote for the current user on a post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts"
                ],
                "summary": "Removes a vote for the current user on a post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The songID of the posted song",
                        "name": "songID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The user who posted the post spotify ID",
                        "name": "posterSpotifyID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.CreatePostDTO": {
            "type": "object",
            "properties": {
                "rating": {
                    "type": "integer"
                },
                "songID": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "requests.UpdatePostRequestDTO": {
            "type": "object",
            "properties": {
                "rating": {
                    "type": "integer"
                },
                "review": {
                    "type": "string"
                }
            }
        },
...
        "responses.Comment": {
            "type": "object",
            "properties": {
                "commentID": {
                    "type": "integer"
                },
                "commentText": {
                    "type": "string"
                },
                "commentorID": {
                    "type": "string"
                },
                "commentorUsername": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "dislikes": {
                    "type": "integer"
                },
                "likes": {
                    "type": "integer"
                },
                "postSpotifyID": {
                    "type": "string"
                },
                "songID": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "responses.PaginationResponse-array_responses_Comment-time_Time": {
            "type": "object",
            "properties": {
                "dataResponse": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.Comment"
                    }
                },
                "paginationKey": {
                    "type": "string"
                }
            }
        },
...
        "responses.PostPreview": {
            "type": "object",
            "properties": {
                "albumArtURI": {
                    "type": "string"
                },
                "albumID": {
                    "type": "string"
                },
                "albumName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "dislikes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.UserIdentifer"
                    }
                },
                "likes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.UserIdentifer"
                    }
                },
                "rating": {
                    "type": "integer"
                },
                "songID": {
                    "type": "string"
                },
                "songName": {
                    "type": "string"
                },
                "spotifyID": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "responses.UserIdentifer": {
            "type": "object",
            "properties": {
                "spotifyID": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Authorization header value\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}